+++ changelog.md
@@ -33,6 +33,7 @@ Codion Change Log
 - EntityPanel.addDetailPanel() exceptions improved, test added.
 - EntityPanel, EntityTablePanel, configuration instance now defensively copied.
 - EntityTablePanel.setEditComponentFactory() and setTableCellEditorFactory() moved to Config.
+- EntityTablePanel.referentialIntegrityErrorHandling() moved to Config.
 
 ## 0.17.31
 ### is.codion.common